Break table to its own component
