+++ src/applications/debt-letters/components/DebtLettersDownload.jsx
@@ -1,29 +1,11 @@
-import React, { useState } from 'react';
-import orderBy from 'lodash/orderBy';
-import environment from 'platform/utilities/environment';
-import recordEvent from 'platform/monitoring/record-event';
+import React from 'react';
 import { Link } from 'react-router';
-import moment from 'moment';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Breadcrumbs from '@department-of-veterans-affairs/formation-react/Breadcrumbs';
+import { DebtLettersTable } from './DebtLettersTable';
 
 const DebtLettersDownload = ({ debtLinks, isVBMSError }) => {
-  const [sortBy, setSortBy] = useState('date');
-  const [direction, setDirection] = useState('desc');
-
-  const sortedDebtLinks = orderBy(debtLinks, [sortBy], direction);
-
-  const toggleDirection = column => {
-    if (column !== sortBy) {
-      setSortBy(column);
-    }
-    if (direction === 'desc') {
-      return setDirection('asc');
-    }
-    return setDirection('desc');
-  };
-
   const renderAlert = () => (
     <div
       className="usa-alert usa-alert-error vads-u-margin-top--0 vads-u-padding--3"
@@ -51,13 +33,6 @@ const DebtLettersDownload = ({ debtLinks, isVBMSError }) => {
     </div>
   );
 
-  const handleDownloadClick = (type, date) => {
-    return recordEvent({
-      event: 'bam-debt-letter-download',
-      'letter-type': type,
-      'letter-received-date': date,
-    });
-  };
   return (
     <div classNam="vads-l-row vads-u-margin-x--neg2p5">
       <Breadcrumbs className="vads-u-font-family--sans">
@@ -78,83 +53,10 @@ const DebtLettersDownload = ({ debtLinks, isVBMSError }) => {
         </p>
         {isVBMSError && renderAlert()}
         {!isVBMSError &&
-          sortedDebtLinks.length > 0 && (
+          debtLinks.length > 0 && (
             <>
               <h3 className="vads-u-margin-bottom--0">Your debt letters</h3>
-              <table className="vads-u-font-family--sans vads-u-margin-top--3 vads-u-margin-bottom--0 responsive">
-                <thead>
-                  <tr>
-                    <th
-                      className="vads-u-border--0 vads-u-padding-left--3"
-                      onClick={() => toggleDirection('date')}
-                    >
-                      Date <i className="fas fa-sort vads-u-margin-left--0p5" />
-                    </th>
-                    <th
-                      className="vads-u-border--0"
-                      onClick={() => toggleDirection('typeDescription')}
-                    >
-                      Type <i className="fas fa-sort vads-u-margin-left--0p5" />
-                    </th>
-                    <th className="vads-u-border--0">Action</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {sortedDebtLinks.map(debtLetter => (
-                    <tr
-                      key={debtLetter.documentId}
-                      className="vads-u-border-top--1px vads-u-border-bottom--1px"
-                    >
-                      <td className="vads-u-border--0 vads-u-padding-left--3">
-                        {moment(debtLetter.receivedAt).format('MMM D, YYYY')}
-                      </td>
-                      <td className="vads-u-border--0">
-                        {debtLetter.typeDescription}
-                      </td>
-                      <td className="vads-u-border--0">
-                        <a
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          onClick={() =>
-                            handleDownloadClick(
-                              debtLetter.typeDescription,
-                              moment(debtLetter.receivedAt).format(
-                                'MMM D, YYYY',
-                              ),
-                            )
-                          }
-                          download={`${
-                            debtLetter.typeDescription
-                          } dated ${moment(debtLetter.receivedAt).format(
-                            'MMM D, YYYY',
-                          )}`}
-                          href={encodeURI(
-                            `${environment.API_URL}/v0/debt_letters/${
-                              debtLetter.documentId
-                            }`,
-                          )}
-                        >
-                          <i
-                            aria-hidden="true"
-                            role="img"
-                            className="fas fa-download vads-u-padding-right--1"
-                          />
-                          <span aria-hidden="true">Download letter </span>
-                          <span className="sr-only">
-                            Download Second Demand Letter dated{' '}
-                            {moment(debtLetter.receivedAt).format(
-                              'MMM D, YYYY',
-                            )}
-                          </span>
-                          <dfn>
-                            <abbr title="Portable Document Format">(PDF)</abbr>
-                          </dfn>
-                        </a>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
+              <DebtLettersTable debtLinks={debtLinks} />
             </>
           )}
         {!isVBMSError &&
